fix(tui): persist sidebar tree expansion across sessions
